process-image/requirements.txt
google-cloud-storage==2.9.0
google-cloud-vision==3.4.0
